requirements.txt
kafka-python
pyspark
tweepy
pymongo
python-dotenv